Comprehensive analyzer-crate review and rewrite of the system test suite.
Adds ~440 system tests across 6 architectures (x86, x64, aarch64, arm, mips32le,
mips32be), all with named per-arch failure surfaces.  Suite tally: 1242 pass,
0 fail, 137 ignored, clippy clean across the workspace.

Highlights:

  - System tests reorganised: binary_tests/test.c -> fixtures/cases/{arithmetic,
    control,memory,calls,floats,abi,stack,builtins,globals,patterns}.c.  All
    six archs build clean under -Wall -Wextra -Werror.

  - per_arch_test! macro emits one #[test] per (arch, function) pair.  Failing
    cases tagged with BUG-N reasons that reference
    docs/superpowers/plans/2026-04-25-analyzer-known-issues.md.

  - Real correctness fixes (each with paired regression test):
      BUG-2  cfg builder allows CONST-space "fall through to next machine
              insn" idiom (MIPS DIV / SLT)
      BUG-3  / BUG-10  extend_if_needed casts Bool inputs via CastToInt
              (MIPS / ARM / float comparisons)
      BUG-13 IntConst payload widened u64 -> u128 (AArch64 large-array constants)
      BUG-21 pattern crate's int_const(v) is now width-aware: int_const(-50)
              matches at U32, U64, U128 IntConsts without per-arch defaults

  - feat(ir): build_int_const now infallible.  Returns NodeOutputId directly
    (was Result<NodeOutputId>).  Cascades through ~365 caller sites.

  - feat(target): MIPS O32 calling convention added; aarch64 float return
    regs restored to ABI-correct q0/q1.

  - feat(analyzer): vn_mask supports 16-byte SIMD container registers.
    read_reg_vn truncates sub-register reads to declared width.
    find_largest_fitting_register signature tightened to Result<Vn>.

  - feat(opt): eval_int_binary/cmp widened to u128/i128 arithmetic.  KnownBits
    masks widened to u128 (with U256 skip-guard).

  - feat(pattern): IntVar::get returns Option<u128>.  int_const(v) accepts
    impl Into<i128> with width-aware compare.

  - 27 BUG entries documented in known-issues.md; this merge closes BUG-2,
    BUG-3/10, BUG-13, BUG-21 and partially closes BUG-9.

  - 17 float arith / conversion / compare tests un-ignored across multiple
    archs after the U128-panic path was removed.

Pre-existing analyzer issues that remain ignored (referenced in known-issues):
  BUG-1  MIPS MULT/HI-LO not lowered
  BUG-4  ARM conditional select Bool/AnyInt
  BUG-5  BranchIndirect opcode unimplemented (intentional per user)
  BUG-6  tail-call elision (test-design issue, not analyzer bug)
  BUG-7  indirect call CFG MemReadErr
  BUG-8  hardware-FPU float chain (some archs)
  BUG-11 float negation as XOR with sign-bit (canonical lowering, test design)
  BUG-12 escape_via_ptr's external_take_ptr call missing
  BUG-14 ARM large_local_array opt panic
  BUG-15 popcount builtin lowering
  BUG-16/17 x86 BSR/BSF Lzcount mapping
  BUG-18/19/20 pattern-matcher edge cases (test design)


# Conflicts:
#	crates/opt/src/constant_fold/eval_int.rs
#	crates/opt/src/stack_store/tests.rs

+++ crates/pattern/src/matcher/bindings.rs
@@ -40,7 +40,7 @@ pub struct Bindings {
 enum BindingEntry {
     Var(Var, NodeOutputId),
     NodeVar(NodeVar, NodeId),
-    Int(IntVar, u64),
+    Int(IntVar, u128),
     Bool(BoolVar, bool),
     Float(FloatVar, u64),
     IntBinaryOp(IntBinaryOpVar, IntBinaryOp),
@@ -109,7 +109,7 @@ macro_rules! decl_bind_get {
 
 decl_bind_get!(Var,             bind_var,             get,                 Var,             NodeOutputId, "the matched `NodeOutputId` (data output edge)");
 decl_bind_get!(NodeVar,         bind_node_var,        get_node,            NodeVar,         NodeId,       "the matched `NodeId` (control-level node)");
-decl_bind_get!(Int,             bind_int,             get_int,             IntVar,          u64,          "the integer constant value");
+decl_bind_get!(Int,             bind_int,             get_int,             IntVar,          u128,         "the integer constant value");
 decl_bind_get!(Bool,            bind_bool,            get_bool,            BoolVar,         bool,         "the boolean constant value");
 decl_bind_get!(Float,           bind_float,           get_float_bits,      FloatVar,        u64,          "the float constant IEEE 754 bit pattern");
 decl_bind_get!(IntBinaryOp,     bind_int_binary_op,   get_int_binary_op,   IntBinaryOpVar,  IntBinaryOp,  "the [`IntBinaryOp`] variant");
